==> 4-functional-preprocess-nowindowfeaturesextraction.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
from IPython.display import display, HTML,clear_output
import joblib

def readData_basicExtract():
    purchasedf = pd.read_csv('../input/buyinghabitsofdigikalausers/purchase_history.csv')
    purchasedf["created_at"] = pd.to_datetime(purchasedf["created_at"])
    
    sortedPurchasedf = purchasedf.sort_values(by=['user_id','product_id','created_at'])
    del purchasedf
    print('purchase shape dataframe shape: ',sortedPurchasedf.shape)
    display(HTML(sortedPurchasedf.head().to_html()))
    
    ALLdata = sortedPurchasedf.copy()
    ALLdata['diffDay'] = ALLdata.groupby(['user_id','product_id'],as_index=False)['created_at'].transform(lambda x : x-x.shift(1))
    ALLdata['diffDay'] = ALLdata['diffDay'].dt.days
    firstRecord = pd.Timestamp('2019-01-01')
    ALLdata['GoneDays'] = (ALLdata['created_at'] - firstRecord).dt.days
    ALLdata.reset_index(inplace=True)
    print('purchase df with diffDay shape: ',ALLdata.shape)
    display(HTML(ALLdata.tail(20).to_html()))
    print('-----------------------------------------------------------------')
    target = ALLdata.reset_index().groupby(['user_id','product_id'],as_index=False).agg(created_at=('created_at','max'),
                                                                     diffDay=('diffDay',lambda x:x.iloc[-1]),
                                                                     GoneDays=('GoneDays',lambda x:x.iloc[-1]),
                                                                      index = ('index',lambda x:x.iloc[-1]) )
    print('target(last purchase datetime) shape: ',target.shape)
    display(HTML(target.tail(5).to_html()))
    trainData = ALLdata[~ALLdata['index'].isin(target['index'])]
    print('trainData shape: ',trainData.shape)
    display(HTML(trainData.tail(20).to_html()))
    
    return ALLdata,trainData,target

ALLdata,trainData,target = readData_basicExtract()

target.to_csv('targetFor_trainProductUserdf.csv',index=False)

> trainData is seperated last datatime of purchased as a target but ALLdata has all datetimes.

### ---------------------------------------------------------------------------------------

#### Grouping and add some features.

def grouping_addTarget(dataname,data):
    groupeddf = data.groupby(['user_id','product_id'],as_index=False).agg(allDiffDay =('diffDay',lambda x: x.tolist()),
                                                                     LenOfList =('diffDay','size'),
                                                                    allGoneDay=('GoneDays',lambda x: x.tolist()),
                                                                      LastDate = ('created_at','max'))
    
    
    
    groupeddf['ProductProbPerUser'] = groupeddf.groupby('user_id',as_index=False)['LenOfList'].transform(lambda x:x/np.nansum(x))
    
    productProb = pd.DataFrame(data['product_id'].value_counts()).reset_index()
    productProb.columns=['product_id','ProductProb']
    productProb['ProductProb'] = (productProb['ProductProb']/len(productProb['ProductProb']) ).round(4)
    print('productProb: ',productProb.shape)
    display(HTML(productProb.tail(6).to_html()))

    groupeddf=pd.merge(left=groupeddf,right=productProb,how='left',on='product_id')
    
    print('grouped '+dataname+' shape: ',groupeddf.shape)
    with pd.option_context('display.max_rows', 100, 'display.max_columns', 10,'max_colwidth', None):
        display(HTML(groupeddf.tail(6).to_html()))
    
    print('NA in the grouped dataFrame?')
    display(groupeddf.isna().sum())

    return groupeddf

groupedTrainDataset = grouping_addTarget('trainData',trainData)
groupedALLdata = grouping_addTarget('ALLdata',ALLdata)

### ---------------------------------------------------------------------

### Extract some features for each product and for each user. NoWindow aproach.

def extractToApply(x):
    x['avgDiffDay'] = np.nanmean(x['allDiffDay'])
    x['stdDiffDay'] = np.nanstd(x['allDiffDay'])
    x['maxDiffDay'] = np.nanmax(x['allDiffDay'])
    x['minDiffDay'] = np.nanmin(x['allDiffDay'])
    
    x['maxGoneDay'] = np.nanmax(x['allGoneDay'])
    x['minGoneDay'] = np.nanmin(x['allGoneDay'])
    x['medianGoneDay'] = np.nanmedian(x['allGoneDay'])
    x['stdGoneDay'] = np.nanstd(x['allGoneDay'])
    
    return x

#### For "groupedTrainDataset"

import os
workingDirfiles=[]
for dirname, _, filenames in os.walk('/kaggle/working'):
    for filename in filenames:
        workingDirfiles.append(filename)
workingDirfiles

fileName = 'trainProductUserdf_addedSomeFeatures.csv'
if fileName in workingDirfiles:
    trainProductUserdf = pd.read_csv('/kaggle/working/'+fileName)
else:
    trainProductUserdf = groupedTrainDataset.apply(extractToApply,axis=1)
print(trainProductUserdf.shape)
with pd.option_context('display.max_rows', 100, 'display.max_columns', 10,'max_colwidth', None):
    display(HTML(trainProductUserdf.tail(5).to_html()))

#trainProductUserdf.to_csv('trainProductUserdf_addedSomeFeatures.csv',index=False)

#### For "groupedALLdata"

fileName = 'ALLdataProductUserdf_addedSomeFeatures.csv'
if fileName in workingDirfiles:
    ALLdataProductUserdf = pd.read_csv('/kaggle/working/'+fileName)
else:
    ALLdataProductUserdf = groupedALLdata.apply(extractToApply,axis=1)
print(ALLdataProductUserdf.shape)
with pd.option_context('display.max_rows', 100, 'display.max_columns', 10,'max_colwidth', None):
    display(HTML(ALLdataProductUserdf.tail(5).to_html()))

#ALLdataProductUserdf.to_csv('ALLdataProductUserdf_addedSomeFeatures.csv',index=False)

### ----------------------------------------------------------------------------------

#### convert Gone day of a year to datetime format. start from 2019-01-01

def convert_goneDayTo_datetime(data,GoneDayColumn,startFrom='2019-01-01'):
    return pd.to_datetime(data[GoneDayColumn],unit='D',origin=startFrom).dt.strftime('%Y-%m-%d')

#### Add Dummy prediction by (maxGoneDay+avgDiffDay)

### trainProductUserdf

trainProductUserdf['dummyPrediction_goneDay'] = trainProductUserdf['maxGoneDay']+trainProductUserdf['avgDiffDay']
trainProductUserdf['dummyPrediction_datetime']= convert_goneDayTo_datetime(trainProductUserdf,'dummyPrediction_goneDay')
with pd.option_context('display.max_rows', 100, 'display.max_columns', 10,'max_colwidth', None):
    display(HTML(trainProductUserdf.tail(2).to_html()))

trainProductUserdf.to_csv('trainProductUserdf_addedSomeFeatures_addedDummyPrediction.csv',index=False)

### ALLdataProductUserdf

ALLdataProductUserdf['dummyPrediction_goneDay'] = ALLdataProductUserdf['maxGoneDay']+ALLdataProductUserdf['avgDiffDay']
ALLdataProductUserdf['dummyPrediction_datetime']= convert_goneDayTo_datetime(ALLdataProductUserdf,'dummyPrediction_goneDay')
with pd.option_context('display.max_rows', 100, 'display.max_columns', 10,'max_colwidth', None):
    display(HTML(ALLdataProductUserdf.tail(2).to_html()))

ALLdataProductUserdf.to_csv('ALLdataProductUserdf_addedSomeFeatures_addedDummyPrediction.csv',index=False)

### --------------------------------------------------------------------

Now our task is make a model to predict next datetime that a customer will buy a specific product.\
We have two aproach to feature ectraction -> noWindow and window appraoches.\
After feature extraction -> make a prediction model to predict goneDay from base datetime(2019-01-01).

NoWindow feature extraction is done in this notebook.\
Create window approach for additional features see this paper([link](https://www.researchgate.net/publication/346846529_Using_machine_learning_to_predict_the_next_purchase_date_for_an_individual_retail_customer)),these features will use on window version datetime presiction.

Continue the codes in the next notebooks.
