==> 6-digikala-productuser-dummyprediction-eval.ipynb <==


import numpy as np # linear algebra
import pandas as pd # data processing, CSV file I/O (e.g. pd.read_csv)

df = pd.read_csv('../input/functional-preprocess-nowindowfeaturesextraction/trainProductUserdf_addedSomeFeatures_addedDummyPrediction.csv')
target = pd.read_csv('../input/functional-preprocess-nowindowfeaturesextraction/targetFor_trainProductUserdf.csv')

df = df[['user_id','product_id','dummyPrediction_goneDay','dummyPrediction_datetime']]
df.head()

target = target[['user_id','product_id','GoneDays']]
target.head()

result = pd.merge(left=df,right=target,how='inner',on=['user_id','product_id'])
result.head()

result.shape

from sklearn.metrics import mean_absolute_error,mean_squared_error,r2_score

r2_score(result.GoneDays,result.dummyPrediction_goneDay)

mean_absolute_error(result.GoneDays,result.dummyPrediction_goneDay)

mean_squared_error(result.GoneDays,result.dummyPrediction_goneDay)
