==> 8-eda-featureextraction-section2.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
from IPython.display import display, HTML,clear_output
import joblib

splitted_productUserdf = joblib.load('../input/eda-featureextraction-section1/ProductUserdf.pkl')
splitted_productUserdf.rename(columns={'LenOfList':'LengthBeforeSplitting'},inplace=True)
with pd.option_context('display.max_rows', 100, 'display.max_columns', 10,'max_colwidth', None):
    display(HTML(splitted_productUserdf.head().to_html()))

#testing how to extract features from "splitted_allDiffDay" and "splitted_allGoneDay"
x =splitted_productUserdf.iloc[2].to_frame().T
leftPart = x.iloc[:,:5]
# rightPart -> "splitted_allDiffDay" and "splitted_allGoneDay"
rightPart = x.iloc[:,5:]
with pd.option_context('display.max_rows', 100, 'display.max_columns', 10,'max_colwidth', None):
    print('leftPart:')
    display(leftPart)
    print('------------------------------------------------')
    print('rightPart:')
    display(rightPart)
print('------------------------------------------------')
print('features')
L=[]
for idx,col in rightPart.iterrows():
    for lst in col:
        L.append(pd.DataFrame(lst)) 
display(L[0])
display(L[1])
print('------------------------------------------------')
print('merging and Concatenating')
merg=leftPart.merge(L[0],how='cross')
display(merg)
print('------------------------------------------------')
display(pd.concat([merg,L[1]],axis=1))

from numba import jit


def make_windowFreatures(dataFrame):
    df = pd.DataFrame()
    for idx,row in dataFrame.iterrows():
        leftPart =  pd.DataFrame(row.iloc[:5]).T
        # rightPart -> "splitted_allDiffDay" and "splitted_allGoneDay"
        rightPart = row.iloc[5:]
        L=[]
        for item in rightPart:
            L.append(pd.DataFrame(item)) 

        merg=leftPart.merge(L[0],how='cross')
        df = pd.concat([df, pd.concat([merg,L[1]],axis=1) ],axis=0)
    return df

%%time
#test extracting features
df = splitted_productUserdf.iloc[1:5].copy()
with pd.option_context('display.max_rows', 100, 'display.max_columns', 20,'max_colwidth', None):
    display(df)
print('result: ')
df = make_windowFreatures(df)
with pd.option_context('display.max_rows', 100, 'display.max_columns', 20,'max_colwidth', None):
    display(df)

Applying "make_windowFreatures" function to the all dataset.

wF_extracted_productUserdf = make_windowFreatures(splitted_productUserdf)
with pd.option_context('display.max_rows', 100, 'display.max_columns', 20,'max_colwidth', None):
    display(wF_extracted_productUserdf.head())

joblib.dump(wF_extracted_productUserdf,'wF_extracted_productUserdf.pkl')

wF_extracted_productUserdf.shape

I can't continue this notebook right now. There are some mistakes in the type of feature extraction that should be corrected and the window method is probably not suitable, but because I don't have enough time to continue, I will finish this project here.
In addition, the code written here can be a good source for learning.
